requirements.txt
Flask==2.0.3
flask-restx==1.3.0
python-dotenv==1.0.0
flask-Cors==4.0.1
Werkzeug==2.0.3
psycopg2-binary
